+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.1.0
 
 - Initial version
+- Connect to an FTP Server with Authorization
+- Get current directory
+- Change directory
+- Make directory
+- Delete directory
+- Upload file (in binary mode)